made some changes

+++ KhunPoker/include/nodes/ShowdownNode.h
@@ -6,25 +6,26 @@
 
 using std::vector;
 
-class ShowdownNode : public GameTreeNode
-{
-public:
-    // enum ShowDownResult
-    // {
-    //     NOTTIE,
-    //     TIE
-    // };
-    // GameTreeNodeType getType() override;
-    // ShowdownNode(vector<double> tie_payoffs, vector<vector<double>> player_payoffs, double pot, shared_ptr<GameTreeNode> parent);
-    ShowdownNode(double pot, shared_ptr<GameTreeNode> parent);
+class ShowdownNode : public GameTreeNode {
+    public:
+        // enum ShowDownResult
+        // {
+        //     NOTTIE,
+        //     TIE
+        // };
+        // GameTreeNodeType getType() override;
+        // ShowdownNode(vector<double> tie_payoffs, vector<vector<double>> player_payoffs, double pot, shared_ptr<GameTreeNode> parent);
+        ShowdownNode(float pot, shared_ptr<GameTreeNode> parent);
 
-    vector<float> utility(Player player, vector<float> reach_probs) override;
-    // vector<double> get_payoffs(ShowDownResult result, int winner);
-    // double get_payoffs(ShowDownResult result, int winner, int player);
+        vector<float> utility(Player player, vector<float> reach_probs) override;
+        // vector<double> get_payoffs(ShowDownResult result, int winner);
+        // double get_payoffs(ShowDownResult result, int winner, int player);
 
-    // private:
-    //     vector<double> tie_payoffs;
-    //     vector<vector<double>> player_payoffs;
+        
+
+        // private:
+        //     vector<double> tie_payoffs;
+        //     vector<vector<double>> player_payoffs;
 };
 
 #endif // SHOWDOWNNODE_H
